Misc tweaks/fixes and file descriptor handling updates
* Imagine: Use UniqueFileDescriptor in PosixIO constructor and releaseFd()
* Imagine: Add ApplicationContext::openFileUriFd()
* Imagine: Add MaybeUniqueFileDescriptor and allow FDEventSource to optionally own its fd
* Imagine: Eliminate move constructors and destructors for classes where FDEventSource owns the fd
* Imagine: Add dummy assignment and comparison operators for UnusedType and ConstantType
* Imagine: Convert some Gfx type aliases to use the standard float type and clean up config variables
* Imagine: Add PixelDesc::rgba()
* 2600.emu: Fix inverted red/blue palette when phosphor is enabled

+++ imagine/src/io/ArchiveIO.cc
@@ -42,7 +42,7 @@ static FS::file_type makeEntryType(int type)
 	return file_type::unknown;
 }
 
-static constexpr unsigned bufferedIoSize = 32 * 1024;
+static constexpr size_t bufferedIoSize = 32 * 1024;
 
 struct ArchiveControlBlockWithBuffer : public ArchiveControlBlock
 {
